# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TODO API OpenApi Documentation",
        "description": "TODO API application swagger/openapi documentation",
        "contact": {
            "email": "salihanmridha@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "TODO API Server"
        }
    ],
    "paths": {
        "/tasks": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Store new task",
                "description": "Returns task data",
                "operationId": "storeTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResponseDTO"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "403 Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpForbiddenResponseDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task information",
                "description": "Returns task data",
                "operationId": "getTaskById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskShowResponseDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "404 not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpNotFoundResponseDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "403 Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpForbiddenResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update existing task",
                "description": "Returns updated task data",
                "operationId": "updateTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskUpdateRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskShowResponseDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "404 not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpNotFoundResponseDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentResponseDTO"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "403 Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpForbiddenResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete existing task",
                "description": "Deletes a record",
                "operationId": "deleteTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 202
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "task deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "403 Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpForbiddenResponseDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "404 not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpNotFoundResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "HttpForbiddenResponseDTO": {
                "title": "Http forbidden Response DTO",
                "description": "Http forbidden (403 code) Response DTO",
                "properties": {
                    "status": {
                        "title": "status",
                        "type": "integer",
                        "example": 403
                    },
                    "success": {
                        "title": "success",
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "title": "error",
                        "type": "string",
                        "example": "You dont have permission to take this action"
                    }
                },
                "type": "object"
            },
            "HttpNotFoundResponseDTO": {
                "title": "Http not found Response DTO",
                "description": "Http not found (404 code) Response DTO",
                "properties": {
                    "status": {
                        "title": "status",
                        "type": "integer",
                        "example": 404
                    },
                    "success": {
                        "title": "success",
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "title": "error",
                        "type": "string",
                        "example": "Something not found"
                    }
                },
                "type": "object"
            },
...
            "TaskRequestDTO": {
                "title": "Task Request DTO",
                "description": "Task Request DTO",
                "required": [
                    "name",
                    "todo_id"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of task",
                        "type": "string",
                        "example": "My task name"
                    },
                    "todo_id": {
                        "title": "todo_id",
                        "description": "Todo id",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "TaskResponseDTO": {
                "title": "Task Response DTO",
                "description": "Task response DTO",
                "properties": {
                    "status": {
                        "title": "status",
                        "type": "integer",
                        "example": 200
                    },
                    "success": {
                        "title": "success",
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "title": "data",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "My first todo"
                            },
                            "todo_id": {
                                "type": "integer",
                                "example": 1
                            },
                            "created_at": {
                                "type": "string",
                                "example": "2023-03-02T12:46:12.000000Z"
                            },
                            "updated_at": {
                                "type": "string",
                                "example": "2023-03-02T12:46:12.000000Z"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "TaskShowResponseDTO": {
                "title": "Task show Response DTO",
                "description": "Task show response DTO",
                "properties": {
                    "status": {
                        "title": "status",
                        "type": "integer",
                        "example": 200
                    },
                    "success": {
                        "title": "success",
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "title": "data",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "My first task"
                            },
                            "todo_id": {
                                "type": "integer",
                                "example": 1
                            },
                            "created_at": {
                                "type": "string",
                                "example": "2023-03-02T12:46:12.000000Z"
                            },
                            "updated_at": {
                                "type": "string",
                                "example": "2023-03-02T12:46:12.000000Z"
                            },
                            "todo": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "my todo name"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "my todo description"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "TaskUpdateRequestDTO": {
                "title": "Task update Request DTO",
                "description": "Task update Request DTO",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of task",
                        "type": "string",
                        "example": "My task name"
                    }
                },
                "type": "object"
            },
...
            "UnprocessableContentResponseDTO": {
                "title": "Unprocessable Content Response DTO",
                "description": "Unprocessable Content (422 code) Response DTO",
                "properties": {
                    "status": {
                        "title": "status",
                        "type": "integer",
                        "example": 422
                    },
                    "success": {
                        "title": "success",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "title": "message",
                        "type": "string",
                        "example": "Some requested data doesn't match the condition"
                    },
                    "errors": {
                        "title": "errors",
                        "properties": {
                            "key": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Validation error shows for the key"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}